requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
numpy==2.1.3
scikit-learn==1.5.2
pydantic==2.9.2
pydantic-settings==2.6.1

